Firmware audit fixes: persist 5 rhythm params, clear stale ACF data
SETTINGS_VERSION 69→70: pllWarmupBeats, snapHysteresis, dbEmaAlpha,
dbThreshold, dbDecay were tunable via serial but lost on reboot.
Now persisted in StoredMusicParams with validation and defaults.

Also fixes:
- correlationAtLag[256]: memset full array instead of partial loop
- LinearPropagation: BLINKY_ASSERT on numLeds > 256 (was silent return)
- Water::particleColor: explicit uint16_t promotion for clarity

+++ blinky-things/generators/Water.cpp
@@ -188,8 +188,8 @@ void Water::renderParticle(const Particle* p, PixelMatrix& matrix) {
 uint32_t Water::particleColor(uint8_t intensity) const {
     // White/light cyan drops for visibility against blue background
     uint8_t white = intensity;
-    uint8_t blue = min(255, intensity + 40);
-    uint8_t green = intensity * 3 / 4;
+    uint8_t blue = (uint8_t)min(255, (int)intensity + 40);
+    uint8_t green = (uint8_t)((uint16_t)intensity * 3 / 4);
     return ((uint32_t)white << 16) | ((uint32_t)green << 8) | blue;
 }
